src/utils/monitoring_agent.py
import aiohttp
from typing import Dict, Any

class MonitoringAgent:
    def __init__(self, master_agent):
        self.master_agent = master_agent
        self.slack_token = "YOUR_SLACK_TOKEN"
        self.jira_token = "YOUR_JIRA_TOKEN"

    async def process(self, task_type: str, *args, **kwargs) -> Any:
        if task_type == 'slack_activity':
            return await self.monitor_slack_activity(*args, **kwargs)
        elif task_type == 'jira_updates':
            return await self.monitor_jira_updates(*args, **kwargs)
        else:
            raise ValueError(f"Unknown monitoring task type: {task_type}")

    async def monitor_slack_activity(self, user_id: str) -> Dict[str, Any]:
        # This is a placeholder. In a real implementation, you'd use the Slack API
        async with aiohttp.ClientSession() as session:
            async with session.get(f"https://slack.com/api/users.info?user={user_id}", headers={"Authorization": f"Bearer {self.slack_token}"}) as resp:
                data = await resp.json()
                # Process and return relevant data
                return {"user_id": user_id, "last_activity": data.get("last_activity")}

    async def monitor_jira_updates(self, project_key: str) -> Dict[str, Any]:
        # This is a placeholder. In a real implementation, you'd use the JIRA API
        async with aiohttp.ClientSession() as session:
            async with session.get(f"https://your-domain.atlassian.net/rest/api/3/project/{project_key}", headers={"Authorization": f"Basic {self.jira_token}"}) as resp:
                data = await resp.json()
                # Process and return relevant data
                return {"project_key": project_key, "last_update": data.get("last_update")}